Revert "Factor function testing 297"
