+++ package.json
@@ -20,7 +20,7 @@
     "update_info": "node utils/update_info.js",
     "preview_gif": "node utils/preview_gif.js",
     "generate_metadata": "node utils/generate_metadata.js",
-    "test": "jest"
+    "test":"jest"
   },
   "author": "Daniel Eugene Botha (HashLips)",
   "license": "MIT",
