--- src/functions/cleanDna/index.test.js
@@ -1,9 +1,0 @@
-const getAbstractedIndexes = require('./index');
-const { mockDataInput, mockDataOutput } = require('./inputOutputData.js');
-
-describe('cleanDna cero', () => {
-  test('cleanDna', () => {
-    const resultInputFalse = getAbstractedIndexes(mockDataInput);
-    expect(resultInputFalse).toStrictEqual(mockDataOutput);
-  });
-});
