--- src/functions/cleanDna/inputOutputData.js
@@ -1,4 +1,0 @@
-const mockDataInput = '0:Black#1.png';
-const mockDataOutput = 0;
-module.exports.mockDataInput = mockDataInput;
-module.exports.mockDataOutput = mockDataOutput;
