--- src/functions/cleanName/index.test.js
@@ -1,9 +1,0 @@
-const cleanName = require('./index');
-const { mockDataInput, mockDataOutput } = require('./inputOutputData.js');
-
-describe('cleanName cero', () => {
-  test('cleanName', () => {
-    const resultNameWithoutExtension = cleanName(mockDataInput);
-    expect(resultNameWithoutExtension).toStrictEqual(mockDataOutput);
-  });
-});
