+++ src/functions/filterDNAOptions/inputOutputData.js
@@ -1,6 +1,4 @@
-const mockDataInput =
-  '0:Black#1.png-1:White#50.png-1:Medium#20.png-0:Shapes#100.png-1:Low#40.png-2:Middle#50.png';
-const mockDataOutput =
-  '0:Black#1.png-1:White#50.png-1:Medium#20.png-0:Shapes#100.png-1:Low#40.png-2:Middle#50.png';
+const mockDataInput = "0:Black#1.png-1:White#50.png-1:Medium#20.png-0:Shapes#100.png-1:Low#40.png-2:Middle#50.png";
+const mockDataOutput = "0:Black#1.png-1:White#50.png-1:Medium#20.png-0:Shapes#100.png-1:Low#40.png-2:Middle#50.png";
 module.exports.mockDataInput = mockDataInput;
 module.exports.mockDataOutput = mockDataOutput;
