--- src/functions/getElements/index.js
@@ -1,23 +1,0 @@
-const fs = require('fs');
-const getRarityWeight = require('../getRarityWeight');
-const cleanName = require('../cleanName');
-
-const getElements = (path) => {
-  return fs
-    .readdirSync(path)
-    .filter((item) => !/(^|\/)\.[^\/\.]/g.test(item))
-    .map((i, index) => {
-      if (i.includes('-')) {
-        throw new Error(`layer name can not contain dashes, please fix: ${i}`);
-      }
-      return {
-        id: index,
-        name: cleanName(i),
-        filename: i,
-        path: `${path}${i}`,
-        weight: getRarityWeight(i),
-      };
-    });
-};
-
-module.exports = getElements;
