--- src/functions/getRarityWeight/inputOutputData.js
@@ -1,4 +1,0 @@
-const mockDataInput = 'High#30.png';
-const mockDataOutput = 30;
-module.exports.mockDataInput = mockDataInput;
-module.exports.mockDataOutput = mockDataOutput;
