--- src/functions/layersSetup/index.test.js
@@ -1,9 +1,0 @@
-const layerSetup = require('./index');
-const { mockDataInput, mockDataOutput } = require('./inputOutputData.js');
-
-describe('layerSetup cero', () => {
-  test('layerSetup', () => {
-    const resultObjectFile = layerSetup(mockDataInput);
-    expect(resultObjectFile).toEqual(mockDataOutput);
-  });
-});
