--- src/functions/removeQueryStrings/inputOutputData.js
@@ -1,4 +1,0 @@
-const mockDataInput = '0:Black#1.png?setting=value';
-const mockDataOutput = '0:Black#1.png';
-module.exports.mockDataInput = mockDataInput;
-module.exports.mockDataOutput = mockDataOutput;
