+++ src/main.js
@@ -1,6 +1,6 @@
 const basePath = process.cwd();
 const { NETWORK } = require(`${basePath}/constants/network.js`);
-const fs = require('fs');
+const fs = require("fs");
 const sha1 = require(`${basePath}/node_modules/sha1`);
 const { createCanvas, loadImage } = require(`${basePath}/node_modules/canvas`);
 const buildDir = `${basePath}/build`;
@@ -13,16 +13,7 @@ const shuffle = require(`${basePath}/node_modules/lodash/shuffle`);
 const saveMetaDataSingleFile = require(`${basePath}/src/functions/saveMetaDataSingleFile`);
 const filterDNAOptions = require(`${basePath}/src/functions/filterDNAOptions`);
 const buildSetup = require(`${basePath}/src/functions/buildSetup`);
-const getElements = require(`${basePath}/src/functions/getElements`);
 const constructLayerToDna = require(`${basePath}/src/functions/constructLayerToDna`);
-const layersSetup = require(`${basePath}/src/functions/layersSetup`);
-const { saveImage } = require(`${basePath}/src/functions/saveImage`);
-const drawBackground = require(`${basePath}/src/functions/drawBackground`);
-const {
-  addMetadata,
-  metadataList,
-  attributesList,
-} = require(`${basePath}/src/functions/addMetadata`);
 
 const {
   format,
@@ -39,35 +30,154 @@ const {
   namePrefix,
   network,
   solanaMetadata,
-  gif,
+  gif
 } = require(`${basePath}/src/config.js`);
 const HashlipsGiffer = require(`${basePath}/modules/HashlipsGiffer.js`);
 
 const canvas = createCanvas(format.width, format.height);
-const ctx = canvas.getContext('2d');
+const ctx = canvas.getContext("2d");
 ctx.imageSmoothingEnabled = format.smoothing;
 
-//const metadataList = [];
-//let attributesList = [];
+const metadataList = [];
+let attributesList = [];
 const dnaList = new Set();
-const DNA_DELIMITER = '-';
 
 let hashlipsGiffer = null;
+
+const getRarityWeight = (_str) => {
+  const nameWithoutExtension = _str.slice(0, -4);
+  let nameWithoutWeight = Number(
+    nameWithoutExtension.split(rarityDelimiter).pop()
+  );
+  if (isNaN(nameWithoutWeight)) {
+    nameWithoutWeight = 1;
+  }
+  return nameWithoutWeight;
+};
+
+const cleanName = (_str) => {
+  const nameWithoutExtension = _str.slice(0, -4);
+  const nameWithoutWeight = nameWithoutExtension.split(rarityDelimiter).shift();
+  return nameWithoutWeight;
+};
+
+const getElements = (path) => {
+  return fs
+    .readdirSync(path)
+    .filter((item) => !/(^|\/)\.[^\/\.]/g.test(item))
+    .map((i, index) => {
+      if (i.includes("-")) {
+        throw new Error(`layer name can not contain dashes, please fix: ${i}`);
+      }
+      return {
+        id: index,
+        name: cleanName(i),
+        filename: i,
+        path: `${path}${i}`,
+        weight: getRarityWeight(i)
+      };
+    });
+};
+
+const layersSetup = (layersOrder) => {
+  const layers = layersOrder.map((layerObj, index) => ({
+    id: index,
+    elements: getElements(`${layersDir}/${layerObj.name}/`),
+    name:
+      layerObj.options?.displayName !== undefined
+        ? layerObj.options?.displayName
+        : layerObj.name,
+    blend:
+      layerObj.options?.blend !== undefined
+        ? layerObj.options?.blend
+        : "source-over",
+    opacity:
+      layerObj.options?.opacity !== undefined
+        ? layerObj.options?.opacity
+        : 1,
+    bypassDNA:
+      layerObj.options?.bypassDNA !== undefined
+        ? layerObj.options?.bypassDNA
+        : false
+  }));
+  return layers;
+};
+
+const saveImage = (_editionCount) => {
+  fs.writeFileSync(
+    `${buildDir}/images/${_editionCount}.png`,
+    canvas.toBuffer("image/png")
+  );
+};
+
+const genColor = () => {
+  const hue = Math.floor(Math.random() * 360);
+  const pastel = `hsl(${hue}, 100%, ${background.brightness})`;
+  return pastel;
+};
+
+const drawBackground = () => {
+  ctx.fillStyle = background.static ? background.default : genColor();
+  ctx.fillRect(0, 0, format.width, format.height);
+};
+
+const addMetadata = (_dna, _edition) => {
+  const dateTime = Date.now();
+  let tempMetadata = {
+    name: `${namePrefix} #${_edition}`,
+    description: description,
+    image: `${baseUri}/${_edition}.png`,
+    dna: sha1(_dna),
+    edition: _edition,
+    date: dateTime,
+    ...extraMetadata,
+    attributes: attributesList,
+    compiler: "HashLips Art Engine"
+  };
+  if (network === NETWORK.sol) {
+    tempMetadata = {
+      // Added metadata for solana
+      name: tempMetadata.name,
+      symbol: solanaMetadata.symbol,
+      description: tempMetadata.description,
+      // Added metadata for solana
+      seller_fee_basis_points: solanaMetadata.seller_fee_basis_points,
+      image: `${_edition}.png`,
+      // Added metadata for solana
+      external_url: solanaMetadata.external_url,
+      edition: _edition,
+      ...extraMetadata,
+      attributes: tempMetadata.attributes,
+      properties: {
+        files: [
+          {
+            uri: `${_edition}.png`,
+            type: "image/png"
+          }
+        ],
+        category: "image",
+        creators: solanaMetadata.creators
+      }
+    };
+  }
+  metadataList.push(tempMetadata);
+  attributesList = [];
+};
 
 const addAttributes = (_element) => {
   const selectedElement = _element.layer.selectedElement;
   attributesList.push({
     trait_type: _element.layer.name,
-    value: selectedElement.name,
+    value: selectedElement.name
   });
 };
 
 const loadLayerImg = async (_layer) => {
   try {
     const image = await loadImage(`${_layer.selectedElement.path}`);
-    return { layer: _layer, loadedImage: image };
+    return ({ layer: _layer, loadedImage: image });
   } catch (error) {
-    console.error('Error loading image:', error);
+    console.error("Error loading image:", error);
   }
 };
 
@@ -88,19 +198,19 @@ const drawElement = (_renderObject, _index, _layersLen) => {
         text.xGap,
         text.yGap * (_index + 1),
         text.size
-      )
+    )
     : ctx.drawImage(
-        _renderObject.loadedImage,
-        0,
-        0,
-        format.width,
-        format.height
-      );
+      _renderObject.loadedImage,
+      0,
+      0,
+      format.width,
+      format.height
+    );
 
   addAttributes(_renderObject);
 };
 
-const isDnaUnique = (_DnaList = new Set(), _dna = '') => {
+const isDnaUnique = (_DnaList = new Set(), _dna = "") => {
   const _filteredDNA = filterDNAOptions(_dna);
   return !_DnaList.has(_filteredDNA);
 };
@@ -117,25 +227,17 @@ const startCreating = async () => {
   let failedCount = 0;
   let abstractedIndexes = [];
 
-  const growEditionSizeTo =
-    layerConfigurations[layerConfigurations.length - 1].growEditionSizeTo;
-  const abstractedIndexesStart = getAbstractedIndexes(
-    growEditionSizeTo,
-    startI
-  );
+  const growEditionSizeTo = layerConfigurations[layerConfigurations.length - 1].growEditionSizeTo;
+  const abstractedIndexesStart = getAbstractedIndexes(growEditionSizeTo, startI);
   abstractedIndexes = abstractedIndexes.concat(abstractedIndexesStart);
 
   if (shuffleLayerConfigurations) {
     abstractedIndexes = shuffle(abstractedIndexes);
   }
 
-  debugLogs && console.log('Editions left to create: ', abstractedIndexes);
+  debugLogs && console.log("Editions left to create: ", abstractedIndexes);
 
   while (layerConfigIndex < layerConfigurations.length) {
-    console.log(
-      'layerorder: ',
-      layerConfigurations[layerConfigIndex].layersOrder
-    );
     const layers = layersSetup(
       layerConfigurations[layerConfigIndex].layersOrder
     );
@@ -154,53 +256,52 @@ const startCreating = async () => {
           loadedElements.push(loadLayerImg(layer));
         });
 
-        await Promise.all(loadedElements).then((renderObjectArray) => {
-          debugLogs && console.log('Clearing canvas');
-          ctx.clearRect(0, 0, format.width, format.height);
-          if (gif.export) {
-            hashlipsGiffer = new HashlipsGiffer(
-              canvas,
-              ctx,
+        const renderObjectArray = await Promise.all(loadedElements);
+        debugLogs && console.log("Clearing canvas");
+        ctx.clearRect(0, 0, format.width, format.height);
+        if (gif.export) {
+          hashlipsGiffer = new HashlipsGiffer(
+            canvas,
+            ctx,
               `${buildDir}/gifs/${abstractedIndexes[0]}.gif`,
               gif.repeat,
               gif.quality,
               gif.delay
-            );
-            hashlipsGiffer.start();
-          }
-          if (background.generate) {
-            drawBackground();
-          }
-          renderObjectArray.forEach((renderObject, index) => {
-            drawElement(
-              renderObject,
-              index,
-              layerConfigurations[layerConfigIndex].layersOrder.length
-            );
-            if (gif.export) {
-              hashlipsGiffer.add();
-            }
-          });
+          );
+          hashlipsGiffer.start();
+        }
+        if (background.generate) {
+          drawBackground();
+        }
+        renderObjectArray.forEach((renderObject, index) => {
+          drawElement(
+            renderObject,
+            index,
+            layerConfigurations[layerConfigIndex].layersOrder.length
+          );
           if (gif.export) {
-            hashlipsGiffer.stop();
+            hashlipsGiffer.add();
           }
-          debugLogs &&
-            console.log('Editions left to create: ', abstractedIndexes);
-          saveImage(abstractedIndexes[0]);
-          addMetadata(newDna, abstractedIndexes[0]);
-          saveMetaDataSingleFile(abstractedIndexes[0], metadataList);
-          console.log(
+        });
+        if (gif.export) {
+          hashlipsGiffer.stop();
+        }
+        debugLogs && console.log("Editions left to create: ", abstractedIndexes);
+        saveImage(abstractedIndexes[0]);
+        addMetadata(newDna, abstractedIndexes[0]);
+        saveMetaDataSingleFile(abstractedIndexes[0], metadataList);
+        console.log(
             `Created edition: ${abstractedIndexes[0]}, with DNA: ${sha1(
               newDna
             )}`
-          );
-        });
+        );
+
         dnaList.add(filterDNAOptions(newDna));
 
         editionCount++;
         abstractedIndexes.shift();
       } else {
-        console.log('DNA exists!');
+        console.log("DNA exists!");
         failedCount++;
         if (failedCount >= uniqueDnaTorrance) {
           console.log(
